Erros da WishList

+++ Aulas_C/Model/Purchase.cs
@@ -123,7 +123,7 @@ public class Purchase : IValidateDataObject, IDataController<PurchaseDTO,Purchas
                 number_confirmation = this.number_confirmation,
                 number_nf = this.number_nf,
                 payment_type = (int)this.payment_type,
-                purchase_status = (int)this.purchase_status,
+                purchase_status = this.purchase_status,
                 purchase_value = this.purchase_value
             };
